==> notebooks/01_data_exploration.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

# Load Data

RAW_DATA_PATH = "../data/raw/survey_results_public.csv"
RAW_DATA_PATH_2020 = "../data/raw/2020/survey_results_public.csv"

raw_df = pd.read_csv(RAW_DATA_PATH_2020)
raw_df

# Get the data from a random Respondent

raw_df.sample(1).T

raw_df.shape

raw_df.info() 


# This is an alternative of the previous cell. 
# The aim: calculate the number of NULL values
raw_df.isnull().sum().sort_values(ascending=False)

# Calculate the portion of null values
raw_df.isnull().sum().sort_values(ascending=False) / raw_df.shape[0]

# Check the datatypes 
raw_df.info(verbose=True)

NUM_FEATS = ["Age", "YearsCode", "YearsCodePro", "OrgSize"]
raw_df[NUM_FEATS].sample().iloc[0]

INV_FEATS = ['YearsCode', 'YearsCodePro'] # Features to check (why they datatype is object. They are supposed to be numerical)
for inv_feat in INV_FEATS:
    unique_values = raw_df[inv_feat].unique()
    print(f".. {inv_feat}: {unique_values}")
    print(f"--------------------------------")


for feat in NUM_FEATS:
    unique_values = raw_df[feat].unique()
    print(f".. {feat}: {unique_values}")
    print(f"--------------------------------")

for feat in raw_df.columns.values: 
    if "Worked" in feat:
        print(feat)
        
        

FEATS = ["DatabaseWorkedWith", 
         "LanguageWorkedWith", 
         "MiscTechWorkedWith", 
         "NEWCollabToolsWorkedWith", 
         "PlatformWorkedWith", 
         "WebframeWorkedWith"
         ]

LABEL = ["DevType"]

raw_df[FEATS + LABEL].sample(1).iloc[0]

raw_df['DevType'].unique()

# The portion of students in the dataset 
# Student is not a job. We will delete this DevType 

students_df = raw_df[raw_df['DevType'] == 'Student']

print("Portion of students in the dataset: ")
print(f".. {students_df.shape[0] / raw_df.shape[0] * 100:.2f}%")

FEATS = ["DatabaseWorkedWith", 
         "LanguageWorkedWith", 
         "MiscTechWorkedWith", 
         "NEWCollabToolsWorkedWith", 
         "PlatformWorkedWith", 
         "WebframeWorkedWith"
         ]

LABEL = ["DevType"]

OTHER_FEATS = []
for feature in raw_df.columns:
    if feature not in FEATS + LABEL:
        OTHER_FEATS.append(feature)
        
raw_df[OTHER_FEATS].sample(1).iloc[0]


unique_main_branch = raw_df['MainBranch'].unique()
print(f"unique_main_branch: {unique_main_branch}")

PRO_MAIN_BRANCH = ["I am a developer by profession", 
                   "I am not primarily a developer, but I write code sometimes as part of my work/studies", 
                   "I used to be a developer by profession, but no longer am"
                   ]

portion_sum = 0
for branch in PRO_MAIN_BRANCH:
    print(f"Portion of {branch} in the dataset")
    print(f".. {branch}: {raw_df[raw_df['MainBranch'] == branch].shape[0] / raw_df.shape[0] * 100:.2f}%")
    portion_sum += raw_df[raw_df['MainBranch'] == branch].shape[0] / raw_df.shape[0] * 100
    
print(f"------------------------------------------")
print(f"The sum of these portions is: {portion_sum:.2f}%")
print(f"The portion of people who code as a hobby: {(100 - portion_sum):.2f}%")


These values insure as that the dataset could be used in our business case. 
- We have only 7% of students in the dataset
- Among professionals, we have 76% of developers that write code as part of their job and only 11% of programmers that code as a hobby
These values indicates that the tech stacks will be correlated with the DevType. In addition, they reflect a real relationship.

END OF NOTEBOOK
