==> notebooks/03_data_preprocessing.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

# Load Data

RAW_DATA_PATH = "../data/raw/2020/survey_results_public.csv"

raw_df = pd.read_csv(RAW_DATA_PATH)
raw_df

raw_df.info(verbose=True)

## YearsCodePro

data = raw_df.copy()

age_unique_values = data['YearsCodePro'].unique()
age_unique_values

new_values_dict = {
    "YearsCodePro" : {
        "Less than 1 year" : "0", 
        "More than 50 years" : "51"
    }, 
    "YearsCode" : {
        "Less than 1 year" : "0", 
        "More than 50 years" : "51"
    }
}

data['YearsCodePro'] = data['YearsCodePro'].str.replace("Less than 1 year", 
                                 new_values_dict['YearsCodePro']['Less than 1 year'])
data['YearsCodePro'] = data['YearsCodePro'].str.replace("More than 50 years", 
                                 new_values_dict['YearsCodePro']['More than 50 years'])

################################à

data['YearsCode'] = data['YearsCode'].str.replace("Less than 1 year", 
                                 new_values_dict['YearsCode']['Less than 1 year'])
data['YearsCode'] = data['YearsCode'].str.replace("More than 50 years", 
                                 new_values_dict['YearsCode']['More than 50 years'])



# Sanity Check 

YEARS_CODE = ['YearsCodePro', 'YearsCode']
for feat in YEARS_CODE:
    unique_values = data[feat].unique()
    print(f"{feat}: {unique_values}")
    print(f"------------------------")

## Skillset feature

data.columns.values

SKILLS_FEATS = [
    'LanguageWorkedWith',
     'DatabaseWorkedWith',
     'PlatformWorkedWith',
     'WebframeWorkedWith',
     'MiscTechWorkedWith',
] 

data[SKILLS_FEATS].sample().iloc[0]

mask = data[SKILLS_FEATS[0]].isna()
data[~mask][SKILLS_FEATS[0]] 

for feat in SKILLS_FEATS:
    mask = data[feat].isna()
    temp = data[mask][feat]
    temp = data[~mask][feat].str.split(";")
    

temp

# pd.options.mode.copy_on_write = False

SKILLS_FEATS

# pd.options.mode.copy_on_write = False
# for feat in SKILLS_FEATS:
#     mask = data[feat].isna()
# 
#     # Set data[mask][feat] to an empty list []
#     data[feat][mask] = data[feat][mask].apply(lambda x: [])
# 
#     # Split data[~mask][feat] by ";"
#     data[feat][~mask] = data[feat][~mask].apply(lambda x: x.split(";"))


def split_answer(pd_series, delimiter): 
    """Function to split single answer"""
    return pd_series.str.split(delimiter)

splitted_df = pd.DataFrame()

for feat in SKILLS_FEATS:
    modified_series = split_answer(data[feat], delimiter=";")
    mask_null = modified_series.isnull()
    modified_series.loc[mask_null] = modified_series[mask_null].apply(lambda x: [])
    
    splitted_df = pd.concat([splitted_df, modified_series], axis=1)
# data[SKILLS_FEATS].sample().iloc[0]
splitted_df

splitted_df.sample().iloc[0]

for feat in SKILLS_FEATS:
    data[feat] = splitted_df[feat]
    
data.sample().iloc[0]

# DevType

data['DevType'].unique()

def split_answer(pd_series, delimiter): 
    """Function to split single answer"""
    return pd_series.str.split(delimiter)

splitted_df = pd.DataFrame()


modified_series = split_answer(data['DevType'], delimiter=";")
mask_null = modified_series.isnull()
modified_series.loc[mask_null] = modified_series[mask_null].apply(lambda x: [])

splitted_df = pd.concat([splitted_df, modified_series], axis=1)
# data[SKILLS_FEATS].sample().iloc[0]
splitted_df

splitted_df.iloc[22654].values

data['DevType'] = splitted_df 

data.sample().iloc[0]

# Save data 

data.to_pickle("../data/processed/2020/survey_results_public_processed.pkl")
