==> notebooks/07_baseline_model.ipynb <==
import numpy as np
import pandas as pd
import mlflow
from mlflow.tracking import MlflowClient
from pathlib import Path
import os 
import pickle

from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.multioutput import MultiOutputClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import make_pipeline

from sklearn import metrics
from sklearn.metrics import auc, accuracy_score, confusion_matrix, f1_score, precision_score, recall_score


MLFLOW_TRACKING_URI = '../models/mlruns'
LOG_PATH = "../models/temp/"
LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

MLFLOW_EXPERIMENT_NAME = "skills_jobs_stackoverflow"

DATA_PATH = "../data/processed/clean_data/clean_data.pkl"

    # Create log directories

Path(MLFLOW_TRACKING_URI).mkdir(parents=True, exist_ok=True)
Path(LOG_PATH).mkdir(parents=True, exist_ok=True)


# Load data

# Read Data
raw_df = pd.read_pickle(DATA_PATH)
raw_df

# Balance data

# Check the total samples of roles
jobs_df = raw_df["DevType"].copy()
jobs_df.sum(axis=0)

jobs_df.sum(axis=0).describe()['25%']

# Resample roles
samples_per_class = 1200
resampled_roles = []

for role_col in jobs_df.columns:
    sub_df = jobs_df.loc[jobs_df[role_col] == 1].copy()
    
    if len(sub_df) < samples_per_class:
        # Upsample
        sub_df = sub_df.sample(samples_per_class, replace=True, random_state=0)
    else:
        # Downsample
        sub_df = sub_df.sample(samples_per_class, random_state=0) 
    
    resampled_roles.append(sub_df)

# Construct dfs
jobs_df  = pd.concat(resampled_roles)
raw_df = raw_df.loc[jobs_df.index].copy()
raw_df

jobs_df.sum(axis=0)


jobs_df.sum(axis=0).describe()

# Split data into Train and Test Data



X_train, X_test, Y_train, Y_test = train_test_split(raw_df.drop("DevType", axis=1), 
                                                    raw_df["DevType"], 
                                                    random_state=0)

# Train the baseline model 

# mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
# client = MlflowClient()
# mlflow.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)
# exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)

client = MlflowClient()
mlflow.set_experiment(MLFLOW_EXPERIMENT_NAME)

exp = client.get_experiment_by_name(MLFLOW_EXPERIMENT_NAME)

## Logistic regression

# CREATE THE BASELINE MODEL AS A SIMPLE LOGISTIC REGRESSION

baseline_model = make_pipeline(StandardScaler(),
                    MultiOutputClassifier(LogisticRegression()))

baseline_model.fit(X_train.values, Y_train.values)
# predictions =  pd.DataFrame(baseline_model.predict(X_train.values),
#                             columns=Y_train.columns)

def calculate_quality(ground_truth, predictions, metric_function, sort_values=False):
    quality_scores = {}
    for col in predictions.columns:
        role_pred  = predictions[col].copy()
        role_truth = ground_truth[col].copy()
        quality_scores[col] = round(metric_function(role_truth, role_pred) * 100, 2)
        
    quality_scores = pd.Series(quality_scores.values(), index=quality_scores.keys())
    if sort_values:
        quality_scores = quality_scores.sort_values()
    
    return quality_scores

# Evaluate on training set
predictions =  pd.DataFrame(baseline_model.predict(X_train.values),
                            columns=Y_train.columns)
train_scores = {score.__name__: calculate_quality(Y_train, predictions, score) 
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
train_scores = pd.concat(train_scores,axis=1)
train_scores

# Evaluate on test set
predictions =  pd.DataFrame(baseline_model.predict(X_test.values), columns=Y_test.columns)

test_scores = {score.__name__: calculate_quality(Y_test, predictions, score) 
                for score in [accuracy_score, precision_score, recall_score, f1_score]}
test_scores = pd.concat(test_scores,axis=1)
test_scores

mean_test_scores = test_scores.mean()
mean_test_scores

mean_train_scores = train_scores.mean()
mean_train_scores

print(mean_test_scores)
test_scores

# Log run mlflow

# Data details
data_details = {"data_path": DATA_PATH,
                "training_indices": X_train.index.tolist(),
                "test_indices":     X_test.index.tolist(), 
                "features_names":   X_train.columns.droplevel(0).tolist(),
                "targets_names":    Y_train.columns.tolist()}

with open(os.path.join(LOG_PATH, LOG_DATA_PKL), "wb") as output_file:
    pickle.dump(data_details, output_file)

# Model
model = {"model_description": "Baseline model: Logistic Regression ",
         "model_details": str(baseline_model),
         "model_object": baseline_model} 

with open(os.path.join(LOG_PATH, LOG_MODEL_PKL), "wb") as output_file:
    pickle.dump(model, output_file)

# Performance details
classes_metrics = {"train_scores": train_scores, 
                   "test_scores":  test_scores}

with open(os.path.join(LOG_PATH, LOG_METRICS_PKL), "wb") as output_file:
    pickle.dump(classes_metrics, output_file)

# Start a new run and track 
with mlflow.start_run(experiment_id=exp.experiment_id, run_name=model["model_description"]):
    # Log pickles 
    mlflow.log_artifacts(LOG_PATH)
    
    # Track metrics 
    for metric, score in mean_test_scores.items():
        mlflow.log_metric(metric, score)
