==> notebooks/08_prediction.ipynb <==
import pandas as pd
import mlflow
from mlflow.tracking import MlflowClient

import os 
import sklearn
import pickle
import yaml


MLFLOW_TRACKING_URI = '../models/mlruns'
MLFLOW_RUN_ID = "88dd333d1e6a46029a6910bef05c43f3"

LOG_DATA_PKL    =  "data.pkl"
LOG_MODEL_PKL   =  "model.pkl"
LOG_METRICS_PKL =  "metrics.pkl"

CLUSTERS_YAML_PATH = "../data/processed/feature_engineering/features_skills_clusters_description.yaml"

mlflow.set_tracking_uri(MLFLOW_TRACKING_URI)
client = MlflowClient()

run = mlflow.get_run(MLFLOW_RUN_ID)
artificats_path = run.info.artifact_uri
artificats_path

# Load Model

artificats_path

# remove the work file

artificats_path = artificats_path[8:]

LOG_MODEL_PKL

model_path = os.path.join(artificats_path, LOG_MODEL_PKL)
print(model_path)
with open(model_path, "rb") as f:
    model = pickle.load(f)
# 
# model

model

data_path  = os.path.join(artificats_path, LOG_DATA_PKL)
with open(data_path, 'rb') as handle:
    data = pickle.load(handle)

data.keys()

features_names = pd.Series(data["features_names"])
targets_names  = pd.Series(data['targets_names'])
classifier = model['model_object']

classifier

classifier

# Load features cluster

with open(CLUSTERS_YAML_PATH, "r") as stream:
    clusters_config = yaml.unsafe_load(stream)

clusters_config

# Reformat into data frame
cluster_skill_map = [(cluster_name, cluster_skill)
                   for cluster_name, cluster_skills in clusters_config.items()
                   for cluster_skill in cluster_skills]

cluster_skill_map

clusters_df = pd.DataFrame(cluster_skill_map, columns=["cluster_name", "skill"])
clusters_df

# Make predictions 

sample_skills = ['Python', 'Pandas', 'Numpy', 'Keras']


pd.Series(sample_skills).isin(features_names)

sample_clusters = clusters_df.copy()
sample_clusters["sample_skills"] = sample_clusters["skill"].isin(sample_skills)
sample_clusters

cluster_features = sample_clusters.groupby("cluster_name")["sample_skills"].sum()
cluster_features

# Onehotencode skills

skills_names = features_names[~features_names.isin(cluster_features.index)]
skills_names

ohe_skills = pd.Series(skills_names.isin(sample_skills).astype(int).tolist(), 
                       index=skills_names)
ohe_skills

features = pd.concat([ohe_skills,
                      cluster_features])
features

features = features.loc[features_names]
features

predictions = classifier.predict_proba([features.values])
predictions

positive_probs = [prob[0][1] for prob in predictions]
predictions = pd.Series(positive_probs, 
          index=targets_names).sort_values(ascending=False)
predictions

predictions / predictions.sum(axis=0) * 100

(predictions / predictions.sum(axis=0) * 100).sum()
